Try to Fix Cross Thread error
The UI is monitoring nodes as well as node.Data and node.Metadata for property changes. Background reload changes these properties which tries to update the UI from a background thread.  solution is to suspend property updates, but now the UI is out of sync. and some other weird issues. hmm.